profiling: Function to print the stack trace for profiled functions
This patch also markes the structure holding the stack trace "noinit"
meaning that its contents are preserved over a reboot. Useful for
debugging when a program triggers reboots (except for brown-outs and
power-on reboot). Tested on AVR but should work on other architectures
as well.

The function profiling_stack_trace() will print out the current recorded
stack trace, starting with the outermost frame.

Info about each frame is printed on a line by itself, consisting of
Frame number, address of the caller, address of the callee and the
time in ticks since restart at which that function was entered.

Use addr2line to resolve the address back to a function name and line.
Note that only functions that are instrumented through the
"-finstrument-functions" compiler option will be picked up in this stack
trace.