Sensors Lib: Check for successful sensor initialization and print errors
On startup, the sensors process initializes all sensors of the platform. However,
errors are not checked and simply ignored. This may lead to situations in which
the user is unaware of the inavailability of a sensor.

This patch checks the return value of the configure() function and prints out
an error if the configuration fails