Inga I2C Driver: Fixed infinitive retrying of i2c operations
I2C operations on the AVR controllers may fail if the I2C bus is
not ready (ATmega1284P data sheet page 219). This can for example happen,
if you connect external I2C slaves that are not powered all the time. Retrying
infinitively in those situations is usually not a good idea.

This patch adds a retry limit of 800 tries to each repeated I2C operation
to avoid nodes to hang forever.