Convergence Layer: Support for temporary NACKs
In certain situations, a uDTN node cannot handle a bundle at the moment. For
example when the storage is filled, the bundle has to be rejected to avoid
overwriting existing bundles. However, the peer shall retransmit said bundle
later. Other bundles will never be parsable. Examples are bundles not fitting
within a single CL segment. The peer should be able to learn, that resending
this bundle will not help.

To signal the peer the difference between those two situations, we now support
temporary NACKs. If parsing fails, a temporary NACK is send out.

Unfortunately the CL is already complex enough - sorry for making it even more
complex.