Routing: remove potentially very long call chains
When the routing cleared its neighbour list (for PDS-based discovery) all former
neighbours are signalled to the CL to allow deleting all tickets. For those tickets
the CL tells the routing, that those bundles could not be delivered. The routing
would then tell discovery, that the respective neighbours are "down" and discovery
would tell the CL to delete the corresponding tickets.

This cycle is now broken.