uDTN: Add support for dtn apps
DTN apps are applications that are part of uDTN and can be added to your project
by simply adding them to your makefile. This is convinent, since basic DTN system
services such as time synchronization and dtnping echo do not have to be part of
the uDTN core but can be easily included by the user.

To create an App just do the following:
1. Create a directory for your app in /apps. The name for this directory is also
   the name of your app.
2. Put your files into the app directory and create a makefile called
    Makefile.appname with the following content:
   appname_src = all_your_c_files.c
3. Inside your c file include "dtn_apps.h" and define the following struct:
   const struct dtn_app appname = {
    "Name of your App",
    init_function_of_your_app,
   };
3. Within any project or example add the following line to the Makefile to make
   use of your app:
   DTN_APPS += appname