Storage Flash: Initial checkin of a new storage module based on the on-board flash of nodes
This module stores one bundle per flash page and tries to minimize the meta data that is kept
in memory to allow storing as many bundles as possible. Persitent storage is ensured since the
module scans the flash for bundles on startup of the node. This module is significantly faster
than coffee-based stored and should be preferred for use-cases in which the flash is exclusively
used for bundle storage.

A flash_arch.h file is required for each platform to specify the interface of the flash.