Storage: Changed interface of delete function, added format function
The storage delete function returned uint16_t for historic reasons which was
unnecessary since only return values -1, 0, 1 are allowed. This commit changes
the return type to uint8_t.

Furthermore, flash-based storages need some way of formatting the underlying
flash to ensure a safe state. This commit adds the format(void) function that
will initialize the state of the flash.