CL: Option to enable or disable segmentation support for bundles (default enabled)
Actually, support for segmentation of bundles has a significant penalty in terms
of RAM and ROM. On modern platforms such as INGA this is not a problem. However,
the oldish TMote Sky is at the limit anyway and segmentation pushes it over the
cliff. This patch allows to configure if segmentation is supported. The option
is enabled by default.

The size difference on INGA is significant:
     No Segmentation  |  Segmentation
ROM:     53141 bytes  |   55301 bytes
RAM:      7673 bytes  |    7801 bytes