CL: Improved timing behaviour of the Convergence Layer
When a packet could not be send since the radio medium was busy, the behavior
was to wait 0.5s and then retry. However, when a broadcast could not be sent
this actually is not very interesting to us.

The new behaviour uses a timer of 0.1s whenever a unicast packet could not
be sent.