[cooja] Reorganization of memory handling layers
`AddressMemory` is an abstract class for memory access based on address and
size.
`VarMemory` is an abstract class for memory access based on variable names.

On Top of this two memory classes exist:

- `MoteMemory` is the glue between the abstract Memory classes and the
  memory implentation (`MemoryInterface`) of the respective mote emulator.

- `SectionMoteMemory` has the same function as `MoteMemory` but can
  handle different memory sections attached to it.

This minimizes the interface between a mote implementation and the cooja
core, reduces duplicate code and adds further security and debugging
benefits.

Note: Reorganization still incomplete at this point. SectionMoteMemory
i.a. will undergo further modifications.