[cooja] Moved native/contiki address offset handling to ContikiMoteType level
This keeps the rest of the memory implementation independent
from the native platform and allows us to see memory access
debug output in contiki address space just as it will be shown
in the motes own debug printouts.

Additionally, the address type was changed to long at some remaining
code lines allowing us to see the addresses in correct, non-negative
form now.