[cooja] Introduce sensor interface support.
This commit introduces a generic sensor interface
for cooja that can show a motes sensors and allows
to feed them with data.

Wrapper examples for mspsim and a generic wrapper
class for avrora development version are included.

The `Sensor` class is the cooja level implementation
of any sensor. Its includes a reference to the
specific Mote and to a SensorAdapter.

The `SensorAdapter` class is a minimal wrapper interface
to concrete sensor implementations such as provided by
mspsim or avrora.

Each `Sensor` provides one or multiple data `Channel`s.
For example, an accelerometer provides 3 channels,
'x', 'y', and 'z' while an accelerometer provides only a
single channel `temperature`.
Each channel can have different values and a different
`DataModel`. All values are represented by doubles.
The data model specifies the input values the sensor accepts.
E.g., a boolean data model would allow only 0.0 (false)
and 1.0 (true).
For invalid data input a `Strategy` can be selected that determines
how to handle it. Options are
(1) ignore, (2) log waring, (3) throw exception, (4) tune to
closest valid data

A sensor can be fed from different types of feeders
that are all subclasses of `AbstractSensorFeeder`.
Currently implemented feeders are:
- `ManualSensorFeeder`: Allows to set a fixed value
- `RandomSensorFeeder`: Allows to set a range from which normal distributed
  random data is generated
- `GaussianSensorFeeder`: Allows to set mean and deviation from whom
  gaussian distributed random data is generated
- `FileSensorFeeder`: Allows to feed sensors from a csv file with time stamp
  and data columns.

Sensors are displayed in a mote interface visualizer and feeders can be
set up trough a visual interface.
Quick help texts try to make everything as self-explanatory as possible.

All sensor settings integrate in coojas xml hierarchy and
can be saved to (and restored from) .csc files.